Putting title of happy hour in item.
