+++ app/src/main/java/com/jordangellatly/hermosahappyhour/model/Restaurant.kt
@@ -246,6 +246,30 @@ val hennesseysSportEvent = Event(
     image = URI("https://tower12hb.com/wp-content/uploads/2022/08/TR-Sat-Sun-4x6-1-696x1024.jpg")
 )
 
+val mondayNightFootball = Event(
+    title = "Monday Night Football",
+    hours = "5PM - 8PM",
+    specials = listOf(
+        Deal(
+            description = "Bloody Mary's Well",
+            price = "$7"
+        ),
+        Deal(
+            description = "Bloody Mary's Titos",
+            price = "$9"
+        ),
+        Deal(
+            description = "22 oz double shot Skyy or Epsolon cocktails",
+            price = "$16 (add Red Bull for $3.50)"
+        ),
+        Deal(
+            description = "22 oz Mavericks Mimosas schooners",
+            price = "$10"
+        )
+    ),
+    image = URI("https://tower12hb.com/wp-content/uploads/2022/08/TR-Sat-Sun-4x6-1-696x1024.jpg")
+)
+
 val sampleSearchRestaurantData = listOf(
     Restaurant(
         id = 1,
