+++ app/src/main/java/com/jordangellatly/hermosahappyhour/model/RestaurantCollection.kt
@@ -10,7 +10,7 @@ data class RestaurantCollection(
     val type: CollectionType = CollectionType.Normal
 )
 
-enum class CollectionType { Normal, Highlight, Featured }
+enum class CollectionType { Normal, HappyHour, Event }
 
 object RestaurantRepo {
     fun getRestaurants(): List<RestaurantCollection> = restaurantCollections
@@ -20,34 +20,34 @@ object RestaurantRepo {
 private val featuredDeals = RestaurantCollection(
     id = 1L,
     name = "Featured Deals",
-    type = CollectionType.Highlight,
+    type = CollectionType.HappyHour,
     restaurants = sampleSearchRestaurantData
 )
 
 private val featuredEvents = RestaurantCollection(
     id = 2L,
     name = "Featured Events",
-    type = CollectionType.Highlight,
+    type = CollectionType.Event,
     restaurants = sampleSearchRestaurantData
 )
 
-private val popularRestaurants = RestaurantCollection(
-    id = 3L,
-    name = "Popular Restaurants",
-    type = CollectionType.Normal,
-    restaurants = sampleSearchRestaurantData
-)
-
-private val newlyAdded = RestaurantCollection(
-    id = 4L,
-    name = "Newly Added",
-    type = CollectionType.Normal,
-    restaurants = sampleSearchRestaurantData
-)
+//private val popularRestaurants = RestaurantCollection(
+//    id = 3L,
+//    name = "Popular Restaurants",
+//    type = CollectionType.Normal,
+//    restaurants = sampleSearchRestaurantData
+//)
+//
+//private val newlyAdded = RestaurantCollection(
+//    id = 4L,
+//    name = "Newly Added",
+//    type = CollectionType.Normal,
+//    restaurants = sampleSearchRestaurantData
+//)
 
 val restaurantCollections = listOf(
     featuredDeals,
     featuredEvents,
-    popularRestaurants,
-    newlyAdded
+//    popularRestaurants,
+//    newlyAdded
 )
