+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/home/RestaurantCard.kt
@@ -43,7 +43,7 @@ val HighlightCardWidth = 170.dp
 val HighlightCardPadding = 16.dp
 
 // The Cards show a gradient which spans 3 cards and scrolls with parallax.
-private val gradientWidth
+val gradientWidth
     @Composable
     get() = with(LocalDensity.current) {
         (3 * (HighlightCardWidth + HighlightCardPadding).toPx())
@@ -91,7 +91,7 @@ fun RestaurantCollection(
 //            CollectionType.Featured -> {
 //                FeaturedRestaurants(index, restaurantCollection.restaurants, onRestaurantClick)
 //            }
-            CollectionType.Highlight -> {
+            CollectionType.HappyHour -> {
                 HighlightedRestaurants(index, restaurantCollection.restaurants, onRestaurantClick)
             }
             CollectionType.Normal -> {
@@ -109,57 +109,10 @@ fun RestaurantCollectionPreview() {
             id = 1,
             name = "Featured Deals",
             restaurants = sampleSearchRestaurantData,
-            type = CollectionType.Featured
+            type = CollectionType.HappyHour
         )
         RestaurantCollection(
             restaurantCollection = restaurantCollection,
-            onRestaurantClick = {}
-        )
-    }
-}
-
-@Composable
-private fun FeaturedRestaurants(
-    index: Int,
-    restaurants: List<Restaurant>,
-    onRestaurantClick: (Long) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val scroll = rememberScrollState(0)
-    val gradient = when ((index / 2) % 2) {
-        0 -> HermosaHappyHourTheme.colors.gradient6_1
-        else -> HermosaHappyHourTheme.colors.gradient6_2
-    }
-    // The Cards show a gradient which spans 3 cards and scrolls with parallax.
-    val gradientWidth = with(LocalDensity.current) {
-        (6 * (HighlightCardWidth + HighlightCardPadding).toPx())
-    }
-    LazyRow(
-        modifier = modifier,
-        horizontalArrangement = Arrangement.spacedBy(16.dp),
-        contentPadding = PaddingValues(start = 24.dp, end = 24.dp)
-    ) {
-        itemsIndexed(restaurants) { index, snack ->
-            FeaturedDealsItem(
-                snack,
-                onRestaurantClick,
-                index,
-                gradient,
-                gradientWidth,
-                scroll.value
-            )
-        }
-    }
-}
-
-@Preview
-@Composable
-fun FeaturedRestaurantsPreview() {
-    HermosaHappyHourTheme {
-        val restaurants = sampleSearchRestaurantData
-        FeaturedRestaurants(
-            index = 0,
-            restaurants = restaurants,
             onRestaurantClick = {}
         )
     }
@@ -186,9 +139,9 @@ private fun HighlightedRestaurants(
         horizontalArrangement = Arrangement.spacedBy(16.dp),
         contentPadding = PaddingValues(start = 24.dp, end = 24.dp)
     ) {
-        itemsIndexed(restaurants) { index, snack ->
+        itemsIndexed(restaurants) { index, restaurant ->
             HighlightRestaurantItem(
-                snack,
+                restaurant,
                 onRestaurantClick,
                 index,
                 gradient,
@@ -208,117 +161,6 @@ fun HighlightedRestaurantsPreview() {
             index = 0,
             restaurants = restaurants,
             onRestaurantClick = {}
-        )
-    }
-}
-
-@Composable
-fun FeaturedDealsItem(
-    restaurant: Restaurant,
-    onRestaurantClick: (Long) -> Unit,
-    index: Int,
-    gradient: List<Color>,
-    gradientWidth: Float,
-    scroll: Int,
-    modifier: Modifier = Modifier
-) {
-    val left = index * with(LocalDensity.current) {
-        (HighlightCardWidth + HighlightCardPadding).toPx()
-    }
-    val date = getCurrentDateTime()
-    val dateInString = date.toString("EEEE").uppercase()
-    val dailyInfo = restaurant.hoursAndSpecials.find { it.dayOfWeek.toString() == dateInString }
-    val happyHourInfo = dailyInfo?.specialEvents?.find { it.title == "Happy Hour" }
-    val happyHours = happyHourInfo?.hours
-    HappyHourCard(
-        modifier = modifier
-            .size(
-                width = 340.dp,
-                height = 400.dp
-            )
-            .padding(bottom = 16.dp)
-    ) {
-        Column(
-            modifier = Modifier
-                .clickable(onClick = { onRestaurantClick(restaurant.id) })
-                .fillMaxSize()
-        ) {
-            Box(
-                modifier = Modifier
-                    .height(160.dp)
-                    .fillMaxWidth()
-            ) {
-                val gradientOffset = left - (scroll / 3f)
-                Box(
-                    modifier = Modifier
-                        .height(100.dp)
-                        .fillMaxWidth()
-                        .offsetGradientBackground(gradient, gradientWidth, gradientOffset)
-                )
-                RestaurantImage(
-                    imageUrl = restaurant.companyLogoUrl,
-                    contentDescription = null,
-                    modifier = Modifier
-                        .size(120.dp)
-                        .align(Alignment.BottomCenter)
-                )
-            }
-            Spacer(modifier = Modifier.height(8.dp))
-            Text(
-                text = happyHourInfo?.title.toString(),
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-                style = MaterialTheme.typography.h6,
-                color = HermosaHappyHourTheme.colors.textSecondary,
-                modifier = Modifier.padding(horizontal = 16.dp)
-            )
-            Spacer(modifier = Modifier.height(4.dp))
-            Text(
-                text = happyHours.toString(),
-                style = MaterialTheme.typography.body1,
-                color = HermosaHappyHourTheme.colors.textHelp,
-                modifier = Modifier.padding(horizontal = 16.dp)
-            )
-            Column {
-                happyHourInfo?.specials?.forEach { deal ->
-                    Spacer(modifier = Modifier.height(4.dp))
-                    Row(
-                        horizontalArrangement = Arrangement.spacedBy(20.dp),
-                        modifier = Modifier.fillMaxWidth()
-                    ) {
-                        Text(
-                            text = deal.price,
-                            style = MaterialTheme.typography.caption,
-                            color = HermosaHappyHourTheme.colors.brand,
-                            modifier = Modifier
-                                .padding(horizontal = 16.dp)
-                                .width(80.dp)
-                        )
-                        Text(
-                            text = deal.description,
-                            style = MaterialTheme.typography.caption,
-                            color = HermosaHappyHourTheme.colors.textSecondary,
-                            modifier = Modifier.padding(horizontal = 16.dp)
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-fun FeaturedDealsItemPreview() {
-    HermosaHappyHourTheme {
-        val restaurant = sampleSearchRestaurantData.first()
-        FeaturedDealsItem(
-            restaurant = restaurant,
-            onRestaurantClick = {},
-            index = 0,
-            gradient = HermosaHappyHourTheme.colors.gradient6_1,
-            gradientWidth = gradientWidth,
-            scroll = 0
         )
     }
 }
